Integrate HE-300 benchmark as tab in main site
- Create unified tabbed interface with HE-300 as default tab
- Add HE300Content component for embedded benchmark functionality
- Convert /he300 to redirect to main page
- Update site title to "CIRIS.AI Alignment Node"
- Remove fragmented hash-based navigation in favor of state-based tabs
